+++ src/ByteDev.Hibp/HibpResponseFactory.cs
@@ -3,8 +3,8 @@ using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using ByteDev.Hibp.Contract.Response;
 using ByteDev.Hibp.Http;
-using ByteDev.Hibp.Response;
 
 namespace ByteDev.Hibp
 {