+++ src/ByteDev.Hibp/HibpUriFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
+using ByteDev.Hibp.Contract.Request;
 using ByteDev.Hibp.Http;
-using ByteDev.Hibp.Request;
 
 namespace ByteDev.Hibp
 {